added icon tests and cleaning up
